BUGFIX: Fixes the quality change slider failure.
This is a Bugfix for the quality slider bug, a bug where changing the
slider results in a fatal error, preventing batcon from exporting the
images.
The error was, that the changed value was never parsed into an int,
resulting in a type error when the value is read by the Converter.
Solution was to wrap the default config in a parseInt(value, 10)
function call.